# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FundCast API
  description: |
    ## AI-First Social Funding + Forecasting Platform for SaaS Founders
    
    FundCast is the premier platform connecting SaaS founders through prediction markets and social funding mechanisms. 
    Our enterprise-grade API provides comprehensive access to:
    
    - **🔮 Prediction Markets**: Binary, categorical, and scalar markets with dual-engine trading
    - **💰 Social Funding**: Reg CF and Rule 506(c) compliant investment workflows  
    - **🤖 AI Intelligence**: Semantic search, market analytics, and founder matching
    - **👥 Founder Network**: Elite community features with Purple Tier home screen featuring
    - **🛡️ Enterprise Security**: OWASP ASVS Level 2+ compliance with Red Team AI defense
    
    ### Authentication
    
    All API endpoints require authentication via JWT tokens:
    
    ```bash
    Authorization: Bearer <jwt_token>
    ```
    
    ### Rate Limiting
    
    - **Free Tier**: 60 requests/minute
    - **Oracle Tier**: 300 requests/minute  
    - **Whale Tier**: 1000 requests/minute
    - **Purple Tier**: 5000 requests/minute
    - **Kingmaker Tier**: Unlimited
    
    ### API Versioning
    
    We use URL path versioning with semantic versioning principles:
    - **Current**: `/api/v1/`
    - **Beta**: `/api/v2-beta/` (Preview features)
    
    ### Error Handling
    
    All errors follow RFC 7807 Problem Details format:
    
    ```json
    {
...
      "status": 422,
...
    }
    ```
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /subscriptions/checkout:
    post:
      tags: [Subscriptions]
      summary: Create checkout session
      description: Create a new subscription checkout session with LemonSqueezy
      operationId: createCheckoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /subscriptions/my-subscription:
    get:
      tags: [Subscriptions]
      summary: Get current subscription
      description: Retrieve current user's subscription details and status
      operationId: getCurrentSubscription
      responses:
        '200':
          description: Subscription details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionDetails'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # =============================================================================
  # Purple Tier Featuring
  # =============================================================================
...
  /subscriptions/purple-featuring/queue:
    get:
      tags: [Purple Tier]
      summary: Get featuring queue
      description: Get user's Purple tier featuring queue and schedule
      operationId: getFeaturingQueue
      responses:
        '200':
          description: Featuring queue retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturingQueueResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # =============================================================================
  # Markets
  # =============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    SubscriptionTier:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          enum: [Founder, Oracle, Whale, Purple, Kingmaker]
        slug:
          type: string
          enum: [founder, oracle, whale, purple, kingmaker]
        description:
          type: string
        monthly_price_cents:
          type: integer
          description: Price in cents for monthly billing
        annual_price_cents:
          type: integer
          description: Price in cents for annual billing (usually discounted)
        features:
          type: array
          items:
            type: string
          description: List of feature slugs included in this tier
        is_popular:
          type: boolean
          description: Whether this tier should be highlighted as popular
        is_recommended:
          type: boolean
          description: Whether this tier should be highlighted as recommended
        max_markets:
          type: integer
          description: Maximum number of markets user can participate in
        rate_limit:
          type: integer
          description: API rate limit per minute
    CheckoutRequest:
      type: object
      required: [tier_slug, billing_cycle]
      properties:
        tier_slug:
          type: string
          enum: [oracle, whale, purple, kingmaker]
        billing_cycle:
          type: string
          enum: [monthly, annual]
          default: monthly
        trial_days:
          type: integer
          minimum: 0
          maximum: 30
          default: 0
        referral_code:
          type: string
          description: Optional referral code for discounts
    CheckoutResponse:
      type: object
      properties:
        success:
          type: boolean
        checkout_url:
          type: string
          format: uri
          description: LemonSqueezy checkout URL
        tier_info:
          $ref: '#/components/schemas/SubscriptionTier'
        billing_cycle:
          type: string
          enum: [monthly, annual]
        trial_days:
          type: integer
        expires_at:
          type: string
          format: date-time
          description: When the checkout session expires
    SubscriptionDetails:
      type: object
      properties:
        subscription:
          $ref: '#/components/schemas/SubscriptionInfo'
        is_subscribed:
          type: boolean
        can_upgrade:
          type: boolean
        can_cancel:
          type: boolean
        next_billing_date:
          type: string
          format: date-time
        usage_stats:
          type: object
          properties:
            api_calls_this_month:
              type: integer
            markets_participated:
              type: integer
            featuring_impressions:
              type: integer
    SubscriptionInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tier:
          $ref: '#/components/schemas/SubscriptionTier'
        status:
          type: string
          enum: [active, canceled, past_due, paused, trialing]
        billing_cycle:
          type: string
          enum: [monthly, annual]
        current_period_start:
          type: string
          format: date-time
        current_period_end:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    # =============================================================================
    # Purple Tier Featuring Schemas
    # =============================================================================
...
    FeaturingQueueResponse:
      type: object
      properties:
        current_featuring:
          $ref: '#/components/schemas/FeaturingSchedule'
        upcoming_slots:
          type: array
          items:
            $ref: '#/components/schemas/FeaturingSchedule'
        queue_position:
          type: integer
          description: Position in the featuring queue
        estimated_next_featuring:
          type: string
          format: date-time
    FeaturingSchedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        featuring_type:
          type: string
          enum: [hero, grid, story]
        scheduled_start:
          type: string
          format: date-time
        scheduled_end:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, active, completed, cancelled]
        custom_content:
          type: object
          properties:
            custom_bio:
              type: string
            achievement_highlight:
              type: string
            cta_text:
              type: string

    # =============================================================================
    # Market Schemas
    # =============================================================================
...
    Error:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type
        title:
          type: string
          description: Short, human-readable summary of the problem
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation of the problem
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
  responses:
...
    UnauthorizedError:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://fundcast.ai/errors/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Valid authentication credentials are required"
            instance: "/api/v1/users/profile"
    ForbiddenError:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://fundcast.ai/errors/forbidden"
            title: "Forbidden" 
            status: 403
            detail: "This feature requires Purple tier subscription or higher"
            instance: "/api/v1/subscriptions/purple-featuring/queue"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            type: "https://fundcast.ai/errors/validation-failed"
            title: "Validation Failed"
            status: 422
            detail: "The request contains invalid parameters"
            instance: "/api/v1/auth/register"
            errors:
              - field: "password"
                message: "Password must contain at least one uppercase letter"
                code: "password_requirements"
...